content like

+++ src/main/java/com/alan10607/leaf/model/ContLike.java
@@ -19,7 +19,7 @@ public class ContLike {
     private String id;
 
     @Id
-    private Long no;
+    private Integer no;
 
     @Id
     private String userId;
